pyproject.toml
[tool.poetry]
name = "ai-assistant"
version = "0.1.0"
description = "AI assistant service based on Telegram bot"
authors = ["dsbulgakovv"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
hydra-core = "^1.3.2"


[tool.poetry.group.bot.dependencies]
aiogram = "^3.8.0"
asyncio = "^3.4.3"
redis = "^6.0.0"
aiogram-dialog = "^2.3.1"
pytz = "^2025.2"


[tool.poetry.group.voice_to_text.dependencies]
fastapi = "^0.111.0"
asyncio = "^3.4.3"
openai-whisper = "^20240930"


[tool.poetry.group.large_lang_model.dependencies]
torch = "^2.3.1"
fastapi = "^0.111.0"
transformers = "^4.42.3"


[tool.poetry.group.db.dependencies]
sqlalchemy = "^2.0.40"
pandas = "^2.2.3"
asyncpg = "^0.30.0"
fastapi = "^0.111.0"
asyncio = "^3.4.3"
psycopg2-binary = "^2.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
